Restructure e2e tests for future cases
