+++ test/e2e_broker/broker_redelivery_test.go
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-package e2e
+package e2e_broker
 
 import (
 	"context"
