+++ test/e2e_sink/kafka_sink.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package e2e
+package e2e_sink
 
 import (
 	"context"
